+++ dutil-basic/src/main/java/com/dwarfeng/dutil/basic/cna/model/ModelUtil.java
@@ -4808,7 +4808,7 @@ public final class ModelUtil {
 
     }
 
-    // 禁止外部实例化。
     private ModelUtil() {
+        throw new IllegalStateException(DwarfUtil.getExceptionString(ExceptionStringKey.MODELUTIL_9));
     }
 }